requirements.txt
chromadb
openai
python-dotenv
pandas
fastapi 
uvicorn 
jinja2
python-multipart